fix: update rusqlite 0.39 and quick-xml 0.38 with breaking API changes
- rusqlite 0.39 removed FromSql impl for u64/usize; use i64 instead
- quick-xml 0.38 removed BytesText::unescape(); use from_utf8_lossy

+++ apps/desktop/src-tauri/src/commands.rs
@@ -410,7 +410,7 @@ pub struct TableRow {
 pub struct TableQueryResult {
     pub columns: Vec<String>,
     pub rows: Vec<TableRow>,
-    pub total_rows: usize,
+    pub total_rows: i64,
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
@@ -466,7 +466,7 @@ pub async fn query_table(
     }
 
     // Count total rows
-    let total_rows: usize = conn
+    let total_rows: i64 = conn
         .query_row(
             &format!("SELECT COUNT(*) FROM \"{}\"", table_name),
             [],
